Feat(Lotto) : Step3 로또(2등) 피드백
- 로또 1등 당첨이 1명 이상인 경우에 대한 Integer 에서 Long 타입으로 변경

+++ src/main/java/lotto/controller/LottoController.java
@@ -38,10 +38,10 @@ public class LottoController {
 
         WinningStatistics winningStatistics = new WinningStatistics(lottoResultNumbers, lottoBonusNumber);
 
-        Map<RewardTable, Integer> resultGameStatistics = winningStatistics.resultLottoGame(lottos.getTickets());
+        Map<RewardTable, Long> resultGameStatistics = winningStatistics.resultLottoGame(lottos.getTickets());
 
         Reward reward = new Reward();
-        int totalMatchPrice = reward.sumTotalMatchPrice(resultGameStatistics);
+        Long totalMatchPrice = reward.sumTotalMatchPrice(resultGameStatistics);
         double rateOfReturn = LottoUtils.calRateOfReturn(totalMatchPrice, lottoBuyPrice);
 
         resultView.resultGame(resultGameStatistics, rateOfReturn);
