+++ src/main/java/lotto/domain/WinningStatistics.java
@@ -8,10 +8,10 @@ import java.util.Map;
 
 public class WinningStatistics {
     private static final int MINIMUM_MATCH_COUNT = 3;
-    private static final int BASIC_REWARD_SET_ZERO = 0;
+    private static final Long BASIC_REWARD_SET_ZERO = 0L;
 
     private List<Integer> lottoResultNumbers;
-    private Map<RewardTable, Integer> resultGameStatistics;
+    private Map<RewardTable, Long> resultGameStatistics;
     private int lottoBonusNumber;
 
     public WinningStatistics(List<Integer> lottoResultNumbers, int lottoBonusNumber) {
@@ -21,7 +21,7 @@ public class WinningStatistics {
         this.lottoBonusNumber = lottoBonusNumber;
     }
 
-    public Map<RewardTable, Integer> resultLottoGame(List<LottoTicket> lottoTickets) {
+    public Map<RewardTable, Long> resultLottoGame(List<LottoTicket> lottoTickets) {
 
         for(LottoTicket lottoTicket : lottoTickets) {
             int matchCount = lottoTicket.matchLottoNumberCount(this.lottoResultNumbers);
@@ -34,12 +34,12 @@ public class WinningStatistics {
     private void addWinningStatistics(int matchCount, Boolean isBouns) {
         if (matchCount >= MINIMUM_MATCH_COUNT) {
             this.resultGameStatistics.put(RewardTable.of(matchCount, isBouns)
-                    , this.resultGameStatistics.getOrDefault(RewardTable.of(matchCount, isBouns), 0) + 1);
+                    , this.resultGameStatistics.getOrDefault(RewardTable.of(matchCount, isBouns), 0L) + 1);
         }
     }
 
     private void makeResultMap() {
-        Map<RewardTable, Integer> map = new HashMap<>();
+        Map<RewardTable, Long> map = new HashMap<>();
 
         RewardTable[] rewardTables = RewardTable.values();
