+++ src/test/java/lotto/util/LottosUtilTest.java
@@ -54,7 +54,7 @@ public class LottosUtilTest {
     @DisplayName("로또 당첨 금액과 구입 금액을 전달하여 수익률을 응답한다.")
     @ParameterizedTest
     @CsvSource(value = {"0:5000:0", "5000:5000:1", "15000:30000:0.5"}, delimiterString = ":")
-    public void lottoUtil_정답_게임_수익률_테스트(int totalPrice, int buyPrice, double rate) {
+    public void lottoUtil_정답_게임_수익률_테스트(Long totalPrice, int buyPrice, double rate) {
         assertThat(LottoUtils.calRateOfReturn(totalPrice, buyPrice)).isEqualTo(rate);
     }
